This is a hack to workaround #1180 while we are implementing spillback scheduling.

**In this PR:**
- All tasks go through the global scheduler.
- The global scheduler keeps track of how much resources each local scheduler has available based on the tasks that the global scheduler assigns to the local schedulers
- The global scheduler gets notifications when tasks finish, which contributes to the bookkeeping
- The global scheduler ignores dynamic load information from local schedulers
- The global scheduler only assigns a task to a local scheduler when the local scheduler has enough resource capacity.

**Drawbacks:**
- Note that task submission is much slower now because there is no local scheduling.
- This probably won't work with fault tolerance at the moment
- This probably won't work well with tasks that submit tasks and block on them
- This does not work with **actors** (this can be easily fixed)