@joschu 2018-01-04
Awesome, thanks for doing this, yeah I'd be happy to try it. Wheel would be great, since I'm running experiments on Kube instances, so otherwise I'd have to write an install script.