@joschu 2018-01-04
python 3.6 (anaconda) on linux