@robertnishihara 2018-01-04
@joschu Ok, you should be able to install it with

```
pip install -U https://s3-us-west-2.amazonaws.com/ray-wheels/extras/centralized-scheduling/ray-0.3.0-cp36-cp36m-linux_x86_64.whl
```

Let me know if you run into any issues or if it does not behave as expected.

Note that this changes the way that regular remote functions are scheduled. It does not affect how actors are placed.