@joschu 2018-01-04
This version gets through 1 iteration about 15% faster than the previous version, but now it seems to have a different issue. In the old version, I'd get all of the jobs running on the head machine while the other machines sit idle, but in the new version, all of the jobs go to the non-head machines while the head machine sits idle.
<img width="1360" alt="new-ray" src="https://user-images.githubusercontent.com/144178/34589829-ccbf9e60-f167-11e7-8d50-ce69ad82a48c.png">
<img width="1424" alt="old-ray" src="https://user-images.githubusercontent.com/144178/34589840-d5d8d82c-f167-11e7-9709-9aecb91dd4e4.png">