@joschu 2018-01-05
Actually slight update to the performance numbers I'm seeing:
New ray, 4 machine: 95.3s
New ray, 1 machine: 229s
Old ray, 4 machine: 193s
Each iteration is an embarrassingly parallel evaluation of 1024 tasks that take between 0 and 10 seconds. So the linear speedup would be ~58 seconds.