@robertnishihara 2018-01-05
Can you print the output of `ray.global_state.client_table()`? I want to see how many CPUs/GPUs Ray thinks each machine has. Are you starting the head node with different flags than the other nodes?