@joschu 2018-01-05
```
{'172.19.1.107': [{'AuxAddress': '172.19.1.107:48844',
   'CPU': 18.0,
   'ClientType': 'local_scheduler',
   'DBClientID': '964e80833f8da67af065455cf4e8a83efd98e5b8',
   'Deleted': False,
   'GPU': 0.0,
   'LocalSchedulerSocketName': '/tmp/scheduler44146121'},
  {'AuxAddress': '172.19.1.107:48844',
   'ClientType': 'plasma_manager',
   'DBClientID': '9abf20c2dd33fc00846f33a6520794386dd34fc8',
   'Deleted': False,
   'manager_socket_name': '/tmp/plasma_manager91911572',
   'store_socket_name': '/tmp/plasma_store53128493'}],
 '172.19.1.126': [{'AuxAddress': '172.19.1.126:52405',
   'CPU': 18.0,
   'ClientType': 'local_scheduler',
   'DBClientID': 'b8a508a903620d7434ad1025879283c689ef6b23',
   'Deleted': False,
   'GPU': 0.0,
   'LocalSchedulerSocketName': '/tmp/scheduler82959634'},
  {'ClientType': 'global_scheduler',
   'DBClientID': '55b0f84b67302a816fa8d8636103798f8b7e6f47',
   'Deleted': False},
  {'AuxAddress': '172.19.1.126:52405',
   'ClientType': 'plasma_manager',
   'DBClientID': '834785efbd1310ff0374de9cf6fc5cfca7a66ae3',
   'Deleted': False,
   'manager_socket_name': '/tmp/plasma_manager91639307',
   'store_socket_name': '/tmp/plasma_store45217882'}],
 '172.19.1.143': [{'AuxAddress': '172.19.1.143:25664',
   'ClientType': 'plasma_manager',
   'DBClientID': '15ea1ae684fd4f3e6aeef52d4f82f45e17b3d174',
   'Deleted': False,
   'manager_socket_name': '/tmp/plasma_manager71756889',
   'store_socket_name': '/tmp/plasma_store28146624'},
  {'AuxAddress': '172.19.1.143:25664',
   'CPU': 18.0,
   'ClientType': 'local_scheduler',
   'DBClientID': '86740db92f9707b0cd7207d64924ce2177e7665d',
   'Deleted': False,
   'GPU': 0.0,
   'LocalSchedulerSocketName': '/tmp/scheduler80754401'}],
 '172.19.1.156': [{'AuxAddress': '172.19.1.156:35904',
   'CPU': 18.0,
   'ClientType': 'local_scheduler',
   'DBClientID': 'b61c32351ebe0e15e918751584cd56afe4ebe742',
   'Deleted': False,
   'GPU': 0.0,
   'LocalSchedulerSocketName': '/tmp/scheduler40187040'},
  {'AuxAddress': '172.19.1.156:35904',
   'ClientType': 'plasma_manager',
   'DBClientID': '9a264d349281970e2d66fed34f49889ae536a53d',
   'Deleted': False,
   'manager_socket_name': '/tmp/plasma_manager23764647',
   'store_socket_name': '/tmp/plasma_store17126165'}]}
```
Head node:

```
        subprocess.check_call(['ray', 'start', '--head', '--num-gpus=' + os.environ['RCALL_NUM_GPU'], 
            '--node-ip-address', ip, '--redis-port', port, '--num-cpus=' + os.environ['RCALL_NUM_CPU']],
```
Other nodes:
```
        subprocess.check_call(['ray', 'start', '--num-gpus=' + os.environ['RCALL_NUM_GPU'], 
            '--redis-address=' + redis_address, '--num-cpus=' + os.environ['RCALL_NUM_CPU']],
            env=nompienv)
```
Previously `--num-cpus` was omitted, and I added it, but it doesn't seem to have an effect.