@robertnishihara 2018-01-05
The behavior you're seeing is very strange, and I haven't been able to reproduce it. E.g., with 18 CPUs on 4 machines, running something like

```python
import numpy as np
import time

@ray.remote                                         
def f():
    time.sleep(np.random.randint(0, 10))     
    return ray.services.get_node_ip_address()

%time l = ray.get([f.remote() for _ in range(1024)])
```

On average, it should take **1024 * 5 / (4 * 18) = 71s**, which is what I'm seeing.

And the load balancing looks reasonable.

```python
[l.count(ls) for ls in set(l)]  # [247, 249, 281, 247]
```

I'll let you know when I have more ideas for things to try.