@joschu 2018-01-05
I'll try the same thing on my setup and see what happens