@joschu 2018-01-05
nitpick: average time is 4.5