@joschu 2018-01-05
I also investigated the possibility that the durations are not random, but rather the long jobs are all at the beginning or end. (Which is the case for my actual use case.) The load balancing is a bit odd, but the timing is still near-optimal, so the culprit of the previous behavior remains a mystery.
```
import ray, numpy as np, time

@ray.remote
def f(sleepdur):
    time.sleep(sleepdur)
    return ray.services.get_node_ip_address()

def main():
    n = 256
    ncpu = 4 * 18
    sleepdurs = np.linspace(0, 10, n)
    from rcall.extras import ray_setup
    ray_setup.ray_setup()
    tstart = time.time()
    l = ray.get([f.remote(sd) for sd in sleepdurs])
    print('counts:',[l.count(ls) for ls in set(l)])  # [247, 249, 281, 247]
    dur = time.time() - tstart
    print('duration', dur)
    print('duration per job', dur * ncpu / n)
```

Orig (short first):
counts: [72, 69, 57, 58]
duration 20.01033616065979
duration per job 5.627907045185566

Reversed (long first):
counts: [98, 50, 54, 54]
duration 22.361101150512695
duration per job 6.289059698581696

Anyway, notwithstanding the under-used head node, the performance is still much better than before for me. Thanks so much Robert for creating this patched version for me!