@robertnishihara 2018-01-05
You're welcome!

So the head node **is** utilized in these synthetic examples, right? So there's something different about the previous example I guess.