@robertnishihara 2018-01-05
One thing I forgot to mention is that this code currently won't work if you use actors. If you're using actors let me know and I'll update it.